+++ src/AlgebraicRewriting.jl
@@ -1,18 +1,12 @@
 module AlgebraicRewriting
 using Reexport
 
-include("FinSets.jl")
-include("CSets.jl")
-include("StructuredCospans.jl")
-include("PartialMap.jl")
-include("Rewrite.jl")
+include("categorical_algebra/CategoricalAlgebra.jl")
+include("rewrite/Rewrite.jl")
 include("Schedules.jl")
 include("Visuals.jl")
 
-@reexport using .FinSets
-@reexport using .CSets
-@reexport using .StructuredCospans
-@reexport using .PartialMap
+@reexport using .CategoricalAlgebra
 @reexport using .Rewrite
 @reexport using .Schedules
 @reexport using .Visuals
