+++ src/categorical_algebra/CSets.jl
