+++ src/categorical_algebra/FinSets.jl
