+++ src/categorical_algebra/PartialMap.jl
