+++ src/categorical_algebra/StructuredCospans.jl
