+++ src/rewrite/Rewrite.jl
@@ -1,0 +1,18 @@
+module Rewrite
+
+using Reexport
+
+include("RewriteDataStructures.jl")
+include("RewriteUtils.jl")
+include("DPO.jl")
+include("SPO.jl")
+include("SqPO.jl")
+
+@reexport using .RewriteDataStructures
+@reexport using .RewriteUtils
+@reexport using .DPO
+@reexport using .SPO
+@reexport using .SqPO
+
+
+end # module
